fix build errors by separating client pages
